+++ app/models/category.rb
@@ -1,23 +1,22 @@
 class Category < ActiveHash::Base
+  field :accept_to_update, default: true
+  field :segments, default: { short: "SHORT PROGRAM", free: "FREE SKATING"}
   self.data =
     [
      ### senior
      {
        name: "MEN",
-       segments: { short: "SHORT PROGRAM", free: "FREE SKATING" },
        isu_bio_url: "http://www.isuresults.com/bios/fsbiosmen.htm",
        senior: true,
      },
      {
        name: "LADIES",
        isu_bio_url: "http://www.isuresults.com/bios/fsbiosladies.htm",
-       segments: { short: "SHORT PROGRAM", free: "FREE SKATING" },
        senior: true,
      },
      {
        name: "PAIRS",
        isu_bio_url: "http://www.isuresults.com/bios/fsbiospairs.htm",
-       segments: { short: "SHORT PROGRAM", free: "FREE SKATING" },
        senior: true,
      },
      {
@@ -27,22 +26,10 @@ class Category < ActiveHash::Base
        senior: true,
      },
      #### junior
-     {
-       name: "JUNIOR MEN",
-       segments: { short: "SHORT PROGRAM", free: "FREE SKATING" },
-     },
-     {
-       name: "JUNIOR LADIES",
-       segments: { short: "SHORT PROGRAM", free: "FREE SKATING" },
-     },
-     {
-       name: "JUNIOR PAIRS",
-       segments: { short: "SHORT PROGRAM", free: "FREE SKATING" },
-     },
-     {
-       name: "JUNIOR ICE DANCE",
-       segments: { short: "SHORT DANCE", free: "FREE DANCE" },
-     },
+     { name: "JUNIOR MEN"},
+     { name: "JUNIOR LADIES"},
+     { name: "JUNIOR PAIRS"},
+     { name: "JUNIOR ICE DANCE"},
     ]
   class << self
     def senior